+++ openmetadata-ui/src/main/resources/ui/playwright/e2e/Features/TagsSuggestion.spec.ts
@@ -15,8 +15,10 @@ import { TableClass } from '../../support/entity/TableClass';
 import { UserClass } from '../../support/user/UserClass';
 import { performAdminLogin } from '../../utils/admin';
 import { redirectToHomePage } from '../../utils/common';
-import { expandNestedColumn } from '../../utils/nestedColumnUpdatesUtils';
-import { createTableTagsSuggestions } from '../../utils/suggestions';
+import {
+  createTableTagsSuggestions,
+  expandTableSuggestionColumns,
+} from '../../utils/suggestions';
 
 const table = new TableClass();
 const table2 = new TableClass();
@@ -75,19 +77,7 @@ test.describe('Tags Suggestions Table Entity', () => {
       await expect(allAvatarSuggestion).toHaveCount(1);
 
       // Expand nested struct/array columns so their suggestion cards render
-      const cols = table.entityResponseData.columns ?? [];
-      const structCol = cols[2];
-      const arrayCol = (structCol.children ?? [])[1];
-      await expandNestedColumn(
-        page,
-        structCol.fullyQualifiedName ?? '',
-        (structCol.children ?? [])[0].fullyQualifiedName ?? ''
-      );
-      await expandNestedColumn(
-        page,
-        arrayCol.fullyQualifiedName ?? '',
-        (arrayCol.children ?? [])[0].fullyQualifiedName ?? ''
-      );
+      await expandTableSuggestionColumns(page, table);
 
       // Click the first avatar
       await allAvatarSuggestion.nth(0).click();
